Alterado todos id com tipo Long para Integer pois estava dando problema no MySQL

+++ app/models/ProdutoSaida.java
@@ -17,7 +17,7 @@ import java.sql.Date;
 public class ProdutoSaida extends Model{
 
     @Id
-    public long id;
+    public Integer id;
     public int quantidade;
     public double valor;
     @Formats.DateTime(pattern="dd-MM-yyyy")
